initial support for remembering declarations
this patch adds a mixin called implicit-progn-with-declare-mixin and 
uses it instead of implicit-progn-mixin in all the places that allow declares.
It has slot which should contain a list of the declares at the top of the 
implicit-progn.  This patch doesn't do anything clever with the declares, 
it just creates declaration-form objects and points their source slots 
at the original declares, however it would be easy to modify parse-declaration
to generate more usefull declaration objects.