Fix infinite loop in (setf log.level)
argh. we'd defined a method for (setf log.level) specialized on
SYMBOLs, this method called (get-logger symbol) and passed that back
again to (setf log.level). However, when get-logger returned nil
(which happens the symbol didn't specify a logging category) we'd
again call the same method (since nil is a symbol), and not the one
specialized on category objects :(