Fold the AST visiting methods into the form-class definer.
Keeping the method definitions separate is very error-prone
and difficult to maintain. Instead, allow two new attributes
for slots defined by form-class, and use them to generate
the methods. The :copy-with attribute controls wheter the
slot is copied by the copy-form function, and allows one
to specify a non-identity function. The :ast-link attribute
is used to generate the old map/rewrite methods.