Restructure the walker environment object.
The lexical environment contains four distinct namespaces,
so it makes sense to use separate lists for them. This would
also improve performance in case of functions with a lot of
local variables, because function lookups would not have to
enumerate them.

This change also forces a refactoring of the variable and
function walker handlers, which probably fixes some bugs.