Define a macro for capturing a compile-file environment.
Some implementations don't propagate some compile-time
declarations into the global persistent state until
the file is loaded, but instead keep them in a special
compilation environment. Testing interactions with this
environment necessarily involves writing a temporary
file and compiling it.

This operation is nontrivial enough to warrant a macro
extension for with-captured-lexical-environment.